frontend/requirements.txt
click==8.1.7
Flask==2.0.3
gunicorn==20.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==2.0.3
Flask-Login==0.6.3
pip==24.1.2
python-dotenv
requests==2.32.3
